Refactored for multiple materials and dish types. Added wood dishes.

+++ src/minecraft/com/spacechase0/minecraft/decorativestuff/item/StencilItem.java
@@ -2,7 +2,7 @@ package com.spacechase0.minecraft.decorativestuff.item;
 
 public class StencilItem extends SimpleItem
 {
-	public StencilItem( int id, String name, int theStencil )
+	public StencilItem( int id, String name, byte theStencil )
 	{
 		super( id, name + "Stencil" );
 		stencil = theStencil;
@@ -11,16 +11,16 @@ public class StencilItem extends SimpleItem
 		setMaxDamage( 4 );
 	}
 	
-	public int getStencilType()
+	public byte getStencilType()
 	{
 		return stencil;
 	}
 	
-	private final int stencil;
+	private final byte stencil;
 	
-	public static final int NO_STENCIL = 0;
-	public static final int STRAIGHT_STENCIL = 1;
-	public static final int ABSTRACT_STENCIL = 2;
-	public static final int BUBBLE_STENCIL = 3;
-	public static final int STENCIL_COUNT = 3 + 1;
+	public static final byte NO_STENCIL = 0;
+	public static final byte STRAIGHT_STENCIL = 1;
+	public static final byte ABSTRACT_STENCIL = 2;
+	public static final byte BUBBLE_STENCIL = 3;
+	public static final byte STENCIL_COUNT = 3 + 1;
 }
